fixing issue of removing rules for class after each method
the extension does not count with the fact that @AfterClass
at side of the container (RemoteLoadableExtension) is executed
each time when method is started. Each test is executed as "a whole
testsuite" thus you meet @AfterSuite, @AfterClass, @After events
for each method.

Here we changed the logic to removing installed byteman rules
at level of the class not from the container extension
but from the client extension where it's capable to catch
the @ContainerCloseClass event and remove the rules as necessary.